pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evoskill"
version = "0.2.0"
description = "Train-free prompt evolution framework via Textual Gradient Descent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "openai>=1.0,<2.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
evoskill = "evoskill.cli:main"

[tool.setuptools.packages.find]
include = ["evoskill*", "evo_framework*"]
